Added simulation command

+++ cogs/task/__init__.py
@@ -2,7 +2,7 @@ import datetime
 from typing import List, Union
 
 from aioscheduler import TimedScheduler
-from nextcord import Member, PermissionOverwrite, Role, TextChannel
+from nextcord import Member, PermissionOverwrite, Role, TextChannel, slash_command, SlashOption, Interaction
 from nextcord.ext import commands, tasks
 from nextcord.utils import get
 
@@ -30,6 +30,26 @@ class Tasker(commands.Cog):
             "of four days each week to be considered active."
         )
 
+    @slash_command(description="Simulate specific day")
+    @commands.is_owner()
+    async def simulate(
+        self,
+        interaction: Interaction,
+        day: str = SlashOption(
+            name="day",
+            description="The day to simulate",
+            required=True,
+            choices={
+                "Tuesday": "Tuesday",
+                "Friday": "Friday",
+                "Sunday": "Sunday",
+            },
+        ),
+    ):
+        await interaction.response.defer(ephemeral=True)
+        await self.task_functions[day](simulation=True)
+        await interaction.edit_original_message(content="The simulation completed")
+    
     def get_permissions(self, state: bool):
         permission_overwrites = PermissionOverwrite()
         permission_overwrites.send_messages = state
@@ -55,9 +75,10 @@ class Tasker(commands.Cog):
                 target=role, overwrite=self.get_permissions(state=False)
             )
 
-    async def tuesday_task(self):
+    async def tuesday_task(self, simulation: bool = False):
         print("[=] Doing Friday task")
-        await self.bot.prisma.lineup.delete_many()
+        if not simulation:
+            await self.bot.prisma.lineup.delete_many()
 
         for guild in self.bot.guilds:
             if guild.id in Data.IGNORED_GUILDS:
@@ -74,10 +95,10 @@ class Tasker(commands.Cog):
                 for member in SUBMITTED_ROLE.members:
                     await member.remove_roles(SUBMITTED_ROLE)
 
-        if not self.once:
+        if not simulation or not self.once:
             self._day_task("Tuesday")
 
-    async def friday_task(self):
+    async def friday_task(self, simulation: bool = False):
         print("[=] Doing sunday task")
         for guild in self.bot.guilds:
             if guild.id in Data.IGNORED_GUILDS:
@@ -136,10 +157,10 @@ class Tasker(commands.Cog):
 
                     await LINEUPS_CHANNEL.send(content=lineups_message)
 
-        if not self.once:
+        if not simulation or not self.once:
             self._day_task("Friday")
 
-    async def sunday_task(self):
+    async def sunday_task(self, simulation: bool = False):
         print("[=] Doing Monday task")
         prisma: Prisma = self.bot.prisma
         try:
@@ -189,7 +210,7 @@ class Tasker(commands.Cog):
                     f"{COMMISSIONERS_ROLE.mention} {ADMINS_ROLE.mention}"
                 )
 
-        if not self.once:
+        if not simulation or not self.once:
             self._day_task("Sunday")
 
     def _day_task(self, day: Days, hour: int = 17):
